refs #23797; Fixed save of shipping line with additional fees

+++ classes/tabs/EbayFormShippingTab.php
@@ -139,7 +139,7 @@ class EbayFormShippingTab extends EbayTab
                     if ($ship_cost == false) {
                         $ship_cost = 0;
                     }
-                    if (isset($extra_fees) && (int)$ship_cost < $extra_fees[$key]) {
+                    if (isset($extra_fees) && (float)$ship_cost < $extra_fees[$key]) {
                         return false;
                     }
